# backend/docs/swagger.yaml
basePath: /
definitions:
  models.CategoryStat:
    properties:
      count:
        type: integer
      name:
        type: string
    type: object
  models.CreateEquipmentRequest:
    properties:
      asset_code:
        maxLength: 50
        minLength: 1
        type: string
      category:
        maxLength: 50
        type: string
      last_maint_date:
        type: string
      location:
        maxLength: 100
        type: string
      maint_interval:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - asset_code
    - maint_interval
    - name
    type: object
  models.CreateUserRequest:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - staff
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - name
    - password
    - role
    - username
    type: object
  models.DeleteEquipmentRequest:
    properties:
      lid:
        type: string
    required:
    - lid
    type: object
  models.DeleteUserRequest:
    properties:
      lid:
        type: string
    required:
    - lid
    type: object
  models.EquipmentDetail:
    properties:
      asset_code:
        type: string
      category:
        type: string
      last_maint_date:
        type: string
      lid:
        type: string
      location:
        type: string
      maint_interval:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  models.EquipmentPublicResponse:
    properties:
      asset_code:
        type: string
      category:
        type: string
      has_active_report:
        type: boolean
      lid:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  models.EquipmentSummary:
    properties:
      faultRate:
        type: integer
      faulty:
        type: integer
      normal:
        type: integer
      pending:
        type: integer
      repairing:
        type: integer
      total:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.MaintenanceRecord:
    properties:
      asset_code:
        type: string
      created_at:
        type: string
      description:
        type: string
      equipment_id:
        type: string
      equipment_name:
        type: string
      is_resolved:
        type: boolean
      lid:
        type: string
      reporter_type:
        description: '''public'', ''staff'', ''system'''
        type: string
      resolve_note:
        type: string
    type: object
  models.MaintenanceRequest:
    properties:
      description:
        maxLength: 500
        minLength: 1
        type: string
      equipment_id:
        type: string
    required:
    - description
    - equipment_id
    type: object
  models.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  models.MonthlyTrend:
    properties:
      faults:
        type: integer
      maintenance:
        type: integer
      monthKey:
        type: string
      name:
        type: string
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  models.ResolveMaintenanceRequest:
    properties:
      lid:
        type: string
      resolve_note:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - lid
    - resolve_note
    type: object
  models.StatsResponse:
    properties:
      category_stats:
        items:
          $ref: '#/definitions/models.CategoryStat'
        type: array
      equipment_summary:
        $ref: '#/definitions/models.EquipmentSummary'
      monthly_trends:
        items:
          $ref: '#/definitions/models.MonthlyTrend'
        type: array
    type: object
  models.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  models.UpdateEquipmentRequest:
    properties:
      asset_code:
        maxLength: 50
        minLength: 1
        type: string
      category:
        maxLength: 50
        type: string
      lid:
        type: string
      location:
        maxLength: 100
        type: string
      maint_interval:
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - lid
    type: object
  models.UpdateUserRequest:
    properties:
      lid:
        type: string
      name:
        maxLength: 64
        minLength: 1
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - staff
        type: string
    required:
    - lid
    type: object
  models.UserResponse:
    properties:
      lid:
        type: string
      name:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: 這是一個整合 PostgreSQL 與 Gin 的設備管理系統。
  title: 設備管理系統 API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 驗證使用者名稱與密碼，並回傳 Access Token 與 Refresh Token。
      parameters:
      - description: 登入資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登入成功
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 帳號或密碼錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 使用者登入
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: 作廢目前的 Refresh Token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/models.MessageResponse'
      summary: 登出系統
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用 Refresh Token 換取新的 Access Token (支援 token rotation)
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Token 無效或過期
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 刷新 Access Token
      tags:
      - auth
  /api/private/equipment:
    delete:
      consumes:
      - application/json
      description: 刪除一個設備紀錄 (僅限管理員)。若該設備仍有關聯維修紀錄則轉為下架/封存，以保留歷史稽核。
      parameters:
      - description: 刪除設備請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteEquipmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功刪除或封存設備
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 設備不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 刪除或下架設備
      tags:
      - private
    patch:
      consumes:
      - application/json
      description: 修改一個設備紀錄 (僅限管理員)
      parameters:
      - description: 修改設備請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEquipmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修改設備
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 設備不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 資產編號已存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 修改設備
      tags:
      - private
    post:
      consumes:
      - application/json
      description: 建立一個新的設備紀錄 (僅限管理員)
      parameters:
      - description: 新增設備請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateEquipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 成功建立設備
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 資產編號已存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 新增設備
      tags:
      - private
  /api/private/equipments:
    get:
      consumes:
      - application/json
      description: 回傳資料庫中未下架設備的完整資訊，支援分頁與關鍵字搜尋 (僅限管理員與維修人員)
      parameters:
      - description: 每頁筆數 (預設 50，最大 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量 (預設 0)
        in: query
        name: offset
        type: integer
      - description: 關鍵字搜尋 (名稱、資產編號、位置，最大 50 字元)
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: 符合條件的總筆數
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.EquipmentDetail'
            type: array
        "400":
          description: 參數錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 獲取所有設備詳情
      tags:
      - private
  /api/private/maintenance-records:
    get:
      consumes:
      - application/json
      description: 取得所有維修紀錄，支援狀態篩選與分頁 (僅限管理員與維修人員)
      parameters:
      - description: 是否顯示已解決 (true/false)
        in: query
        name: resolved
        type: string
      - description: 每頁筆數 (預設 50，最大 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量 (預設 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: 符合條件的總筆數
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.MaintenanceRecord'
            type: array
        "400":
          description: 查詢參數錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取得維修紀錄列表
      tags:
      - private
  /api/private/maintenance-records/resolve:
    patch:
      consumes:
      - application/json
      description: 將維修紀錄標記為已解決，並將設備狀態恢復為 normal。會驗證紀錄是否已經解決避免重複操作。
      parameters:
      - description: 維修完成請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ResolveMaintenanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修復
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 紀錄不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 紀錄已被解決
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 標記維修完成
      tags:
      - private
  /api/private/stats:
    get:
      consumes:
      - application/json
      description: 取得儀表板與數據分析頁面所需的設備總數、狀態佔比、分類維修數與近六個月趨勢 (管理員與維修人員皆可存取)
      produces:
      - application/json
      responses:
        "200":
          description: 統計數據
          schema:
            $ref: '#/definitions/models.StatsResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取得系統統計聚合數據
      tags:
      - private
  /api/private/user:
    delete:
      consumes:
      - application/json
      description: 刪除使用者 (僅限管理員)。不允許刪除自己或刪除最後一位管理員。
      parameters:
      - description: 刪除使用者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功刪除使用者
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: 不可刪除自己或最後一位管理員
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 使用者不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 刪除使用者
      tags:
      - private
    patch:
      consumes:
      - application/json
      description: 修改使用者資料 (僅限管理員)。不允許自降身分或移除最後一位管理員。
      parameters:
      - description: 修改使用者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功修改使用者
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: 無權限修改自己角色或降級最後管理員
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 使用者不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 修改使用者資料
      tags:
      - private
    post:
      consumes:
      - application/json
      description: 新增一個使用者 (僅限管理員)
      parameters:
      - description: 新增使用者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 成功新增使用者
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 使用者名稱已存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 新增使用者
      tags:
      - private
  /api/private/users:
    get:
      consumes:
      - application/json
      description: 取得所有使用者資料，支援分頁 (僅限管理員)
      parameters:
      - description: 每頁筆數 (預設 50，最大 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量 (預設 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 使用者列表
          headers:
            X-Total-Count:
              description: 符合條件的總筆數
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 查看使用者列表
      tags:
      - private
  /api/public/equipment:
    get:
      consumes:
      - application/json
      description: 透過 asset_code 取得特定設備的資訊
      parameters:
      - description: 設備資產編號
        in: query
        name: asset_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功回傳設備資訊
          schema:
            $ref: '#/definitions/models.EquipmentPublicResponse'
        "400":
          description: 資產編號必填
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 查詢不到設備
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 查詢設備詳情
      tags:
      - public
  /api/public/report:
    post:
      consumes:
      - application/json
      description: 建立一個新的報修紀錄，並將設備狀態更新為 faulty。由 server 端固定 reporter_type 為 public
        並限制請求大小。
      parameters:
      - description: 報修資訊
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MaintenanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功提交報修紀錄
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: 請求格式錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 設備不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: 設備已有未處理紀錄
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 伺服器內部錯誤
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 提交報修紀錄
      tags:
      - public
securityDefinitions:
  BearerAuth:
    description: 請輸入 "Bearer <Your_JWT_Token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"